First test written and some fixes in the SecantMethod class

+++ Src/Math/Numerical/SecantMethodApp.h
@@ -36,6 +36,6 @@ public:
     SecantMethodApp();
     ~SecantMethodApp();
 
-    static double compute(Polynomial&, double, double, double, double, bool&, int&);
+    static double compute(Polynomial&, double, double, double, int, bool&, int&);
 
 };
